feat: remove last remnants of SecretManager
